feat: inject working badly
