+++ .codespellrc
@@ -1,0 +1,3 @@
+[codespell]
+ignore-words-list =
+  covert
