+++ .github/workflows/ci.yml
@@ -14,12 +14,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v6
-      - name: Load tool versions
-        run: cat .github/versions.env >> "$GITHUB_ENV"
-      - name: Install codespell
-        run: pip install --user "codespell==${{ env.CODESPELL_VERSION }}"
-      - name: Run codespell
-        run: codespell -I .codespellignore
+      - uses: codespell-project/actions-codespell@v2
 
   golangci-lint:
     runs-on: ubuntu-latest
@@ -126,10 +121,10 @@ jobs:
           name: dist-linux-${{ matrix.goarch }}
           path: dist/
       - name: Set up Docker Buildx
-        if: matrix.test == 'inject'
+        if: contains(matrix.test, 'inject')
         uses: docker/setup-buildx-action@v3
       - name: Build test docker image
-        if: matrix.test == 'inject'
+        if: contains(matrix.test, 'inject')
         uses: docker/build-push-action@v6
         with:
           context: test/docker
@@ -141,12 +136,7 @@ jobs:
         shell: bash
         run: |
           EXECUTABLE=$(pwd)/$(find dist/ -path "*linux*${{ matrix.goarch }}*" -name dingopie -type f | head -n 1)
-          if [ -z "$EXECUTABLE" ]; then
-            echo "Error: Could not find dingopie binary in dist/"
-            exit 1
-          fi
-          echo "dingopie binary found at: $EXECUTABLE"
-          chmod +x $EXECUTABLE
+          chmod -v +x $EXECUTABLE
           make ${{ matrix.test }} EXECUTABLE=$EXECUTABLE
 
   test-windows:
@@ -169,6 +159,5 @@ jobs:
         run: |
           $exe = (Resolve-Path .\dist\dingopie_windows_amd64_v1\dingopie.exe).Path
           $exe = $exe -replace '\\', '/'
-          echo "dingopie binary found at: $exe"
           make ${{ matrix.test }} EXECUTABLE="$exe"
         shell: pwsh
