+++ .goreleaser.yml
@@ -1,11 +1,24 @@
 version: 2
 
+project_name: dingopie
+
+before:
+  hooks:
+    - go mod tidy
+
 builds:
   - env:
       - CGO_ENABLED=0
+      # Garble everything except for:
+      # go-packet (takes too long)
+      # go-cmp    (only used in tests)
+      # cobra     (uses reflection for commands)
+      - GOGARBLE=github.com/coreos/go-iptables,github.com/creack/pty,github.com/florianl/go-nfqueue,github.com/nblair2/go-dnp3/v2,github.com/schollz/progressbar/v3,golang.org/x/term
+    # Trick goreleaser into using garble as the build tool
     tool: garble
     command: "-literals"
     flags:
+      # Makefile and ci should both set and record the GARBLE_SEED
       - -seed={{ envOrDefault "GARBLE_SEED" "random" }}
       - build
       - -trimpath
