+++ internal/shell/shell.go
@@ -30,7 +30,7 @@ import (
 	"strconv"
 
 	"github.com/nblair2/dingopie/internal"
-	"github.com/nblair2/go-dnp3/dnp3"
+	"github.com/nblair2/go-dnp3/v2/dnp3"
 	"golang.org/x/term"
 )
